backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.2
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
cors==1.0.1
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
nltk==3.8.1
flask==3.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.0
tensorflow>=2.13.0
pillow>=10.0.0
matplotlib>=3.8.0
langchain==0.1.4
chromadb==0.4.15
sentence-transformers==2.7.0
torch>=2.0.0
pdf2image>=1.16.3
PyMuPDF>=1.23.0
opencv-python>=4.8.0
aiofiles>=23.2.0
httpx>=0.25.0
asyncio>=3.4.3
python-pptx>=0.6.21
fitz>=0.0.1.dev2
faiss-cpu>=1.7.4
langgraph>=0.2.30
neo4j>=5.25.0
rapidfuzz>=3.9.7
apify-client>=1.8.1
